Move user-associated components to rest_auth.users app
