+++ rest_auth/urls.py
@@ -1,11 +1,9 @@
 from collections import OrderedDict
 
-from django.conf.urls import include, url
+from django.conf.urls import url
 from rest_framework.routers import APIRootView
 
-from . import routers
 from .views import (
-    EmailVerificationConfirmView,
     LoginView, LogoutView, PasswordChangeView, PasswordForgotConfirmView,
     PasswordForgotView, PasswordResetDoneView,
 )
@@ -39,10 +37,4 @@ for pattern in urlpatterns:
 urlpatterns += [
     url(r'^api-root/$',
         APIRootView.as_view(api_root_dict=api_root), name='api-root'),
-
-    url(r'^user/', include(routers.router.urls)),
-    url(r'^user/v/'
-        r'(?P<uidb64>[0-9A-Za-z_\-]+)/'
-        r'(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
-        EmailVerificationConfirmView.as_view(), name='verify_email_confirm'),
 ]
